requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
filelock==3.11.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post4
sympy==1.11.1
threadpoolctl==3.1.0
torch==2.0.0
torch-summary==1.4.5
torchaudio==2.0.1
torchmetrics==0.11.4
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
